Moved new traits class to experimental folder.
